Fixed bug #492260: Crash on rotate on x86_64. Patch by Ignacy Gawędzki.
2007-11-01  Michael J. Chudobiak  <mjc@svn.gnome.org>

        * NEWS:
        * libgthumb/jpegutils/jpegtran.c: (update_exif_thumbnail):
        Fixed bug #492260: Crash on rotate on x86_64.
        Patch by Ignacy Gawędzki.


svn path=/branches/gthumb-2-10/; revision=2043

+++ libgthumb/jpegutils/jpegtran.c
@@ -209,7 +209,7 @@ static void
 update_exif_thumbnail (ExifData    *edata,
 		       JXFORM_CODE  transform)
 {
-	unsigned int osize;
+	size_t osize;
 	unsigned char *out;
 
 	if (edata == NULL || edata->data == NULL)
